* Added casts to ensure math is done as longs instead of ints (was getting negative offset exceptions when passed to RandomAccessFile.seek(long))

+++ jdem846/src/us/wthr/jdem846/input/esri/GridAscii.java
@@ -135,7 +135,7 @@ public class GridAscii extends DataSource
 	public void loadRow(int row) throws DataSourceException
 	{
 		this.cachedRow = row;
-		long seekTo = row * (header.getColumns() * 4);
+		long seekTo = (long) row * (long)(header.getColumns() * 4l);
 		cache.load(seekTo);
 	}
